eliminar con modal sin funcionar
